+++ VungleAdapter/src/main/java/com/chartboost/mediation/vungleadapter/VungleAdapter.kt
@@ -1,30 +1,54 @@
 /*
  * Copyright 2023-2024 Chartboost, Inc.
- * 
+ *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE file.
  */
 
 package com.chartboost.mediation.vungleadapter
 
+import android.app.Activity
 import android.content.Context
 import android.util.Size
-import com.chartboost.heliumsdk.domain.*
-import com.chartboost.heliumsdk.utils.PartnerLogController
-import com.chartboost.heliumsdk.utils.PartnerLogController.PartnerAdapterEvents.*
-import com.vungle.ads.AdConfig
-import com.vungle.ads.BannerAd
-import com.vungle.ads.BannerAdListener
-import com.vungle.ads.BannerAdSize
-import com.vungle.ads.BaseAd
-import com.vungle.ads.BaseFullscreenAd
-import com.vungle.ads.FullscreenAdListener
-import com.vungle.ads.InitializationListener
-import com.vungle.ads.InterstitialAd
-import com.vungle.ads.RewardedAd
-import com.vungle.ads.RewardedAdListener
-import com.vungle.ads.VungleAds
-import com.vungle.ads.VungleError
+import com.chartboost.chartboostmediationsdk.ad.ChartboostMediationBannerAdView.ChartboostMediationBannerSize.Companion.asSize
+import com.chartboost.chartboostmediationsdk.domain.*
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.BIDDER_INFO_FETCH_FAILED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.BIDDER_INFO_FETCH_STARTED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.BIDDER_INFO_FETCH_SUCCEEDED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.DID_CLICK
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.DID_DISMISS
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.DID_REWARD
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.DID_TRACK_IMPRESSION
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.GDPR_CONSENT_DENIED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.GDPR_CONSENT_GRANTED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.GDPR_CONSENT_UNKNOWN
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.GDPR_NOT_APPLICABLE
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.INVALIDATE_FAILED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.INVALIDATE_STARTED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.INVALIDATE_SUCCEEDED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.LOAD_FAILED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.LOAD_STARTED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.LOAD_SUCCEEDED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.SETUP_FAILED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.SETUP_STARTED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.SETUP_SUCCEEDED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.SHOW_FAILED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.SHOW_STARTED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.SHOW_SUCCEEDED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.USER_IS_NOT_UNDERAGE
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.USER_IS_UNDERAGE
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.USP_CONSENT_DENIED
+import com.chartboost.chartboostmediationsdk.utils.PartnerLogController.PartnerAdapterEvents.USP_CONSENT_GRANTED
+import com.chartboost.core.consent.ConsentKey
+import com.chartboost.core.consent.ConsentKeys
+import com.chartboost.core.consent.ConsentManagementPlatform
+import com.chartboost.core.consent.ConsentValue
+import com.chartboost.core.consent.ConsentValues
+import com.chartboost.mediation.vungleadapter.VungleAdapterConfiguration.adOrientation
+import com.chartboost.mediation.vungleadapter.VungleAdapterConfiguration.adapterVersion
+import com.chartboost.mediation.vungleadapter.VungleAdapterConfiguration.backBtnImmediatelyEnabled
+import com.vungle.ads.*
 import com.vungle.ads.VungleError.Companion.AD_FAILED_TO_DOWNLOAD
 import com.vungle.ads.VungleError.Companion.ASSET_DOWNLOAD_ERROR
 import com.vungle.ads.VungleError.Companion.INVALID_APP_ID
@@ -34,7 +58,6 @@ import com.vungle.ads.VungleError.Companion.NO_SERVE
 import com.vungle.ads.VungleError.Companion.PLACEMENT_NOT_FOUND
 import com.vungle.ads.VungleError.Companion.SDK_NOT_INITIALIZED
 import com.vungle.ads.VungleError.Companion.SERVER_RETRY_ERROR
-import com.vungle.ads.VunglePrivacySettings
 import kotlinx.coroutines.CancellableContinuation
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.serialization.json.Json
@@ -49,45 +72,6 @@ import kotlin.coroutines.resume
 class VungleAdapter : PartnerAdapter {
     companion object {
         /**
-         * Flag to optionally set to specify whether the back button will be immediately enabled
-         * during the video ad, or it will be inactive until the on screen close button appears (the default behavior).
-         * It can be set at any time and will take effect on the next ad request.
-         *
-         * Once enabled, the Android back button allows the user to skip the video ad and proceed
-         * to the post-roll if one exists. If the ad does not have a post-roll, it simply ends.
-         */
-        var backBtnImmediatelyEnabled = false
-            set(value) {
-                field = value
-                PartnerLogController.log(
-                    CUSTOM,
-                    "Vungle back button setting is ${if (value) "enabled" else "disabled"}.",
-                )
-            }
-
-        /**
-         * If set, Vungle will check if it has an ad that can be rendered in the specified orientation.
-         * This flag can be set at any time and will take effect on the next ad request.
-         *
-         * See [AdConfig.Orientation] for available options.
-         */
-        var adOrientation: Int? = null
-            set(value) {
-                field = value
-                PartnerLogController.log(
-                    CUSTOM,
-                    "Vungle ad orientation set to ${
-                        when (value) {
-                            AdConfig.PORTRAIT -> "PORTRAIT"
-                            AdConfig.LANDSCAPE -> "LANDSCAPE"
-                            AdConfig.AUTO_ROTATE -> "AUTO_ROTATE"
-                            else -> "UNSPECIFIED"
-                        }
-                    }.",
-                )
-            }
-
-        /**
          * Key for parsing the Vungle app ID.
          */
         private const val APP_ID_KEY = "vungle_app_id"
@@ -101,13 +85,13 @@ class VungleAdapter : PartnerAdapter {
          */
         internal fun getChartboostMediationError(vungleError: VungleError?) =
             when (vungleError?.code) {
-                NO_SERVE, AD_FAILED_TO_DOWNLOAD -> ChartboostMediationError.CM_LOAD_FAILURE_NO_FILL
-                SERVER_RETRY_ERROR, ASSET_DOWNLOAD_ERROR -> ChartboostMediationError.CM_AD_SERVER_ERROR
-                NETWORK_ERROR, NETWORK_UNREACHABLE -> ChartboostMediationError.CM_NO_CONNECTIVITY
-                SDK_NOT_INITIALIZED -> ChartboostMediationError.CM_INITIALIZATION_FAILURE_UNKNOWN
-                INVALID_APP_ID -> ChartboostMediationError.CM_INITIALIZATION_FAILURE_INVALID_CREDENTIALS
-                PLACEMENT_NOT_FOUND -> ChartboostMediationError.CM_LOAD_FAILURE_INVALID_PARTNER_PLACEMENT
-                else -> ChartboostMediationError.CM_PARTNER_ERROR
+                NO_SERVE, AD_FAILED_TO_DOWNLOAD -> ChartboostMediationError.LoadError.NoFill
+                SERVER_RETRY_ERROR, ASSET_DOWNLOAD_ERROR -> ChartboostMediationError.OtherError.AdServerError
+                NETWORK_ERROR, NETWORK_UNREACHABLE -> ChartboostMediationError.OtherError.NoConnectivity
+                SDK_NOT_INITIALIZED -> ChartboostMediationError.InitializationError.Unknown
+                INVALID_APP_ID -> ChartboostMediationError.InitializationError.InvalidCredentials
+                PLACEMENT_NOT_FOUND -> ChartboostMediationError.LoadError.InvalidPartnerPlacement
+                else -> ChartboostMediationError.OtherError.PartnerError
             }
 
         /**
@@ -123,37 +107,9 @@ class VungleAdapter : PartnerAdapter {
     }
 
     /**
-     * Get the partner name for internal uses.
+     * The Vungle adapter configuration.
      */
-    override val partnerId: String
-        get() = "vungle"
-
-    /**
-     * Get the partner name for external uses.
-     */
-    override val partnerDisplayName: String
-        get() = "Vungle"
-
-    /**
-     * Get the Vungle SDK version.
-     */
-    override val partnerSdkVersion: String
-        get() = VungleAds.getSdkVersion()
-
-    /**
-     * Get the Vungle adapter version.
-     *
-     * You may version the adapter using any preferred convention, but it is recommended to apply the
-     * following format if the adapter will be published by Chartboost Mediation:
-     *
-     * Chartboost Mediation.Partner.Adapter
-     *
-     * "Chartboost Mediation" represents the Chartboost Mediation SDK’s major version that is compatible with this adapter. This must be 1 digit.
-     * "Partner" represents the partner SDK’s major.minor.patch.x (where x is optional) version that is compatible with this adapter. This can be 3-4 digits.
-     * "Adapter" represents this adapter’s version (starting with 0), which resets to 0 when the partner SDK’s version changes. This must be 1 digit.
-     */
-    override val adapterVersion: String
-        get() = BuildConfig.CHARTBOOST_MEDIATION_VUNGLE_ADAPTER_VERSION
+    override var configuration: PartnerAdapterConfiguration = VungleAdapterConfiguration
 
     /**
      * Initialize the Vungle SDK so that it is ready to request ads.
@@ -164,11 +120,11 @@ class VungleAdapter : PartnerAdapter {
     override suspend fun setUp(
         context: Context,
         partnerConfiguration: PartnerConfiguration,
-    ): Result<Unit> {
+    ): Result<Map<String, Any>> {
         PartnerLogController.log(SETUP_STARTED)
 
         return suspendCancellableCoroutine { continuation ->
-            fun resumeOnce(result: Result<Unit>) {
+            fun resumeOnce(result: Result<Map<String, Any>>) {
                 if (continuation.isActive) {
                     continuation.resume(result)
                 }
@@ -190,7 +146,8 @@ class VungleAdapter : PartnerAdapter {
                                     adapterVersion,
                                 )
 
-                                resumeOnce(Result.success(PartnerLogController.log(SETUP_SUCCEEDED)))
+                                PartnerLogController.log(SETUP_SUCCEEDED)
+                                resumeOnce(Result.success(emptyMap()))
                             }
 
                             override fun onError(vungleError: VungleError) {
@@ -212,7 +169,7 @@ class VungleAdapter : PartnerAdapter {
                 PartnerLogController.log(SETUP_FAILED, "Missing App ID.")
                 resumeOnce(
                     Result.failure(
-                        ChartboostMediationAdException(ChartboostMediationError.CM_INITIALIZATION_FAILURE_INVALID_CREDENTIALS),
+                        ChartboostMediationAdException(ChartboostMediationError.InitializationError.InvalidCredentials),
                     ),
                 )
             }
@@ -220,103 +177,44 @@ class VungleAdapter : PartnerAdapter {
     }
 
     /**
-     * Notify Vungle of the user's CCPA consent status, if applicable.
+     * Notify Vungle if the user is underage.
      *
      * @param context The current [Context].
-     * @param hasGrantedCcpaConsent True if the user has granted CCPA consent, false otherwise.
-     * @param privacyString The CCPA privacy string.
+     * @param isUserUnderage True if the user is underage, false otherwise.
      */
-    override fun setCcpaConsent(
+    override fun setIsUserUnderage(
         context: Context,
-        hasGrantedCcpaConsent: Boolean,
-        privacyString: String,
+        isUserUnderage: Boolean,
     ) {
         PartnerLogController.log(
-            if (hasGrantedCcpaConsent) {
-                CCPA_CONSENT_GRANTED
+            if (isUserUnderage) {
+                USER_IS_UNDERAGE
             } else {
-                CCPA_CONSENT_DENIED
+                USER_IS_NOT_UNDERAGE
             },
         )
 
-        VunglePrivacySettings.setCCPAStatus(hasGrantedCcpaConsent)
-    }
-
-    /**
-     * Notify the Vungle SDK of the GDPR applicability and consent status.
-     *
-     * @param context The current [Context].
-     * @param applies True if GDPR applies, false otherwise.
-     * @param gdprConsentStatus The user's GDPR consent status.
-     */
-    override fun setGdpr(
-        context: Context,
-        applies: Boolean?,
-        gdprConsentStatus: GdprConsentStatus,
-    ) {
-        PartnerLogController.log(
-            when (applies) {
-                true -> GDPR_APPLICABLE
-                false -> GDPR_NOT_APPLICABLE
-                else -> GDPR_UNKNOWN
-            },
-        )
-
-        PartnerLogController.log(
-            when (gdprConsentStatus) {
-                GdprConsentStatus.GDPR_CONSENT_UNKNOWN -> GDPR_CONSENT_UNKNOWN
-                GdprConsentStatus.GDPR_CONSENT_GRANTED -> GDPR_CONSENT_GRANTED
-                GdprConsentStatus.GDPR_CONSENT_DENIED -> GDPR_CONSENT_DENIED
-            },
-        )
-
-        if (applies == true) {
-            VunglePrivacySettings.setGDPRStatus(
-                optIn = gdprConsentStatus == GdprConsentStatus.GDPR_CONSENT_GRANTED,
-                consentMessageVersion = "",
-            )
-        }
-    }
-
-    /**
-     * Notify Vungle of the COPPA subjectivity.
-     *
-     * @param context The current [Context].
-     * @param isSubjectToCoppa True if the user is subject to COPPA, false otherwise.
-     */
-    override fun setUserSubjectToCoppa(
-        context: Context,
-        isSubjectToCoppa: Boolean,
-    ) {
-        PartnerLogController.log(
-            if (isSubjectToCoppa) {
-                COPPA_SUBJECT
-            } else {
-                COPPA_NOT_SUBJECT
-            },
-        )
-
-        // NO-OP. See: https://support.vungle.com/hc/en-us/articles/360047780372#coppa-implementation-0-2
+        VunglePrivacySettings.setCOPPAStatus(isUserUnderage)
     }
 
     /**
      * Get a bid token if network bidding is supported.
      *
      * @param context The current [Context].
-     * @param request The [PreBidRequest] instance containing relevant data for the current bid request.
+     * @param request The [PartnerAdPreBidRequest] instance containing relevant data for the current bid request.
      *
      * @return A Map of biddable token Strings.
      */
     override suspend fun fetchBidderInformation(
         context: Context,
-        request: PreBidRequest,
-    ): Map<String, String> {
+        request: PartnerAdPreBidRequest,
+    ): Result<Map<String, String>> {
         PartnerLogController.log(BIDDER_INFO_FETCH_STARTED)
 
         val token = VungleAds.getBiddingToken(context) ?: ""
 
         PartnerLogController.log(if (token.isNotEmpty()) BIDDER_INFO_FETCH_SUCCEEDED else BIDDER_INFO_FETCH_FAILED)
-        return mapOf("bid_token" to token)
+        return Result.success(mapOf("bid_token" to token))
     }
 
     /**
@@ -335,22 +233,18 @@ class VungleAdapter : PartnerAdapter {
     ): Result<PartnerAd> {
         PartnerLogController.log(LOAD_STARTED)
 
-        return when (request.format.key) {
-            AdFormat.BANNER.key, "adaptive_banner" -> {
+        return when (request.format) {
+            PartnerAdFormats.BANNER -> {
                 loadBannerAd(context, request, partnerAdListener)
             }
 
-            AdFormat.INTERSTITIAL.key, AdFormat.REWARDED.key -> {
+            PartnerAdFormats.INTERSTITIAL, PartnerAdFormats.REWARDED, PartnerAdFormats.REWARDED_INTERSTITIAL -> {
                 loadFullscreenAd(context, request, partnerAdListener)
             }
 
             else -> {
-                if (request.format.key == "rewarded_interstitial") {
-                    loadFullscreenAd(context, request, partnerAdListener)
-                } else {
-                    PartnerLogController.log(LOAD_FAILED)
-                    Result.failure(ChartboostMediationAdException(ChartboostMediationError.CM_LOAD_FAILURE_UNSUPPORTED_AD_FORMAT))
-                }
+                PartnerLogController.log(LOAD_FAILED)
+                Result.failure(ChartboostMediationAdException(ChartboostMediationError.LoadError.UnsupportedAdFormat))
             }
         }
     }
@@ -358,34 +252,32 @@ class VungleAdapter : PartnerAdapter {
     /**
      * Attempt to show the currently loaded Vungle ad.
      *
-     * @param context The current [Context]
+     * @param activity The current [Activity]
      * @param partnerAd The [PartnerAd] object containing the ad to be shown.
      *
      * @return Result.success(PartnerAd) if the ad was successfully shown, Result.failure(Exception) otherwise.
      */
     override suspend fun show(
-        context: Context,
+        activity: Activity,
         partnerAd: PartnerAd,
     ): Result<PartnerAd> {
         PartnerLogController.log(SHOW_STARTED)
 
-        return when (partnerAd.request.format.key) {
+        return when (partnerAd.request.format) {
             // Banner ads do not have a separate "show" mechanism.
-            AdFormat.BANNER.key, "adaptive_banner" -> {
+            PartnerAdFormats.BANNER -> {
                 PartnerLogController.log(SHOW_SUCCEEDED)
                 Result.success(partnerAd)
             }
 
-            AdFormat.INTERSTITIAL.key, AdFormat.REWARDED.key -> showFullscreenAd(partnerAd)
+            PartnerAdFormats.INTERSTITIAL, PartnerAdFormats.REWARDED, PartnerAdFormats.REWARDED_INTERSTITIAL ->
+                showFullscreenAd(partnerAd)
+
             else -> {
-                if (partnerAd.request.format.key == "rewarded_interstitial") {
-                    showFullscreenAd(partnerAd)
-                } else {
-                    PartnerLogController.log(SHOW_FAILED)
-                    Result.failure(
-                        ChartboostMediationAdException(ChartboostMediationError.CM_SHOW_FAILURE_UNSUPPORTED_AD_FORMAT),
-                    )
-                }
+                PartnerLogController.log(SHOW_FAILED)
+                Result.failure(
+                    ChartboostMediationAdException(ChartboostMediationError.ShowError.UnsupportedAdFormat),
+                )
             }
         }
     }
@@ -400,17 +292,67 @@ class VungleAdapter : PartnerAdapter {
     override suspend fun invalidate(partnerAd: PartnerAd): Result<PartnerAd> {
         PartnerLogController.log(INVALIDATE_STARTED)
 
-        return when (partnerAd.request.format.key) {
+        return when (partnerAd.request.format) {
             /**
              * Only invalidate banner ads.
              * For fullscreen ads, since Vungle does not provide an ad in the load callback, we don't
              * have an ad in PartnerAd to invalidate.
              */
-            AdFormat.BANNER.key, "adaptive_banner" -> destroyBannerAd(partnerAd)
+            PartnerAdFormats.BANNER -> destroyBannerAd(partnerAd)
             else -> {
                 PartnerLogController.log(INVALIDATE_SUCCEEDED)
                 Result.success(partnerAd)
             }
+        }
+    }
+
+    override fun setConsents(
+        context: Context,
+        consents: Map<ConsentKey, ConsentValue>,
+        modifiedKeys: Set<ConsentKey>,
+    ) {
+        val consent = consents[configuration.partnerId]?.takeIf { it.isNotBlank() }
+            ?: consents[ConsentKeys.GDPR_CONSENT_GIVEN]?.takeIf { it.isNotBlank() }
+        consent?.let {
+            if (VungleAdapterConfiguration.isGdprStatusOverridden) {
+                return@let
+            }
+            if (it == ConsentValues.DOES_NOT_APPLY) {
+                PartnerLogController.log(GDPR_NOT_APPLICABLE)
+                return@let
+            }
+
+            PartnerLogController.log(
+                when (it) {
+                    ConsentValues.GRANTED -> GDPR_CONSENT_GRANTED
+                    ConsentValues.DENIED -> GDPR_CONSENT_DENIED
+                    else -> GDPR_CONSENT_UNKNOWN
+                },
+            )
+
+            VunglePrivacySettings.setGDPRStatus(
+                optIn = it == ConsentValues.GRANTED,
+                consentMessageVersion = "",
+            )
+        }
+
+        val hasGrantedUspConsent =
+            consents[ConsentKeys.CCPA_OPT_IN]?.takeIf { it.isNotBlank() }
+                ?.equals(ConsentValues.GRANTED)
+                ?: consents[ConsentKeys.USP]?.takeIf { it.isNotBlank() }
+                    ?.let { ConsentManagementPlatform.getUspConsentFromUspString(it) }
+        hasGrantedUspConsent?.let {
+            if (VungleAdapterConfiguration.isCcpaStatusOverridden) {
+                return@let
+            }
+            PartnerLogController.log(
+                when (hasGrantedUspConsent) {
+                    true -> USP_CONSENT_GRANTED
+                    false -> USP_CONSENT_DENIED
+                }
+            )
+
+            VunglePrivacySettings.setCCPAStatus(hasGrantedUspConsent)
         }
     }
 
@@ -461,7 +403,7 @@ class VungleAdapter : PartnerAdapter {
                 BannerAd(
                     context,
                     request.partnerPlacement,
-                    getVungleBannerSize(request.size),
+                    getVungleBannerSize(request.bannerSize?.asSize()),
                 )
 
             vungleBanner.adListener =
@@ -580,30 +522,30 @@ class VungleAdapter : PartnerAdapter {
             }
 
             fun loadVungleFullScreenAd(fullscreenAd: BaseFullscreenAd) {
-                fullscreenAd.adListener = VungleFullScreenAdListener(
-                    request = request,
-                    listener = listener,
-                    continuationRef = continuationRef
-                )
+                fullscreenAd.adListener =
+                    VungleFullScreenAdListener(
+                        request = request,
+                        listener = listener,
+                        continuationRef = continuationRef,
+                    )
                 fullscreenAd.load(adm)
             }
 
             when (request.format) {
-                AdFormat.INTERSTITIAL -> loadVungleFullScreenAd(InterstitialAd(context, request.partnerPlacement, adConfig))
-                AdFormat.REWARDED -> loadVungleFullScreenAd(RewardedAd(context, request.partnerPlacement, adConfig))
+                PartnerAdFormats.INTERSTITIAL -> loadVungleFullScreenAd(InterstitialAd(context, request.partnerPlacement, adConfig))
+                PartnerAdFormats.REWARDED, PartnerAdFormats.REWARDED_INTERSTITIAL ->
+                    loadVungleFullScreenAd(
+                        RewardedAd(context, request.partnerPlacement, adConfig),
+                    )
                 else -> {
-                    if (request.format.key == "rewarded_interstitial") {
-                        loadVungleFullScreenAd(RewardedAd(context, request.partnerPlacement, adConfig))
-                    } else {
-                        PartnerLogController.log(LOAD_FAILED)
-                        resumeOnce(
-                            Result.failure(
-                                ChartboostMediationAdException(
-                                    ChartboostMediationError.CM_LOAD_FAILURE_UNSUPPORTED_AD_FORMAT,
-                                ),
+                    PartnerLogController.log(LOAD_FAILED)
+                    resumeOnce(
+                        Result.failure(
+                            ChartboostMediationAdException(
+                                ChartboostMediationError.LoadError.UnsupportedAdFormat,
                             ),
-                        )
-                    }
+                        ),
+                    )
                 }
             }
         }
@@ -658,7 +600,7 @@ class VungleAdapter : PartnerAdapter {
                     resumeOnce(
                         Result.failure(
                             ChartboostMediationAdException(
-                                ChartboostMediationError.CM_SHOW_FAILURE_AD_NOT_FOUND,
+                                ChartboostMediationError.ShowError.AdNotFound,
                             ),
                         ),
                     )
@@ -672,7 +614,7 @@ class VungleAdapter : PartnerAdapter {
                         resumeOnce(
                             Result.failure(
                                 ChartboostMediationAdException(
-                                    ChartboostMediationError.CM_SHOW_FAILURE_AD_NOT_READY,
+                                    ChartboostMediationError.ShowError.AdNotReady,
                                 ),
                             ),
                         )
@@ -686,7 +628,7 @@ class VungleAdapter : PartnerAdapter {
                     resumeOnce(
                         Result.failure(
                             ChartboostMediationAdException(
-                                ChartboostMediationError.CM_SHOW_FAILURE_WRONG_RESOURCE_TYPE,
+                                ChartboostMediationError.ShowError.WrongResourceType,
                             ),
                         ),
                     )
@@ -709,7 +651,6 @@ class VungleAdapter : PartnerAdapter {
         private val listener: PartnerAdListener,
         private val continuationRef: WeakReference<CancellableContinuation<Result<PartnerAd>>>,
     ) : FullscreenAdListener, RewardedAdListener {
-
         fun resumeOnce(result: Result<PartnerAd>) {
             continuationRef.get()?.let {
                 if (it.isActive) {
@@ -832,7 +773,7 @@ class VungleAdapter : PartnerAdapter {
             Result.success(partnerAd)
         } ?: run {
             PartnerLogController.log(INVALIDATE_FAILED, "Ad is null.")
-            Result.failure(ChartboostMediationAdException(ChartboostMediationError.CM_INVALIDATE_FAILURE_AD_NOT_FOUND))
+            Result.failure(ChartboostMediationAdException(ChartboostMediationError.InvalidateError.AdNotFound))
         }
     }
 }